pyproject.toml
[project]
name = "github-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
